pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=1.2.6",
    "django>=5.1.7",
    "django-cors-headers>=4.7.0",
    "django-ninja>=1.3.0",
    "django-ninja-extra>=0.22.6",
    "fastapi[all]>=0.115.12",
    "google-genai>=1.9.0",
    "google-generativeai>=0.8.4",
    "lancedb>=0.21.2",
    "openai>=1.70.0",
    "pandas>=2.2.3",
    "pgvector>=0.4.0",
    "psycopg>=3.2.6",
    "psycopg2-binary>=2.9.10",
    "pymongo>=4.11.3",
    "pypdf>=5.4.0",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.0",
]
